Another attempt to fix 4760. (cherry picked from commit 4dadce6ebca29e6f6d7120a44541fd99034417f2)

+++ scripts/firewall/vyatta-ipset.pl
@@ -74,7 +74,10 @@ sub ipset_delete_member {
 
     die "Error: undefined member" if ! defined $member; 
     my $group = new Vyatta::IpTables::IpSet($set_name);
-    return $group->delete_member($member);
+    if ($group->exists($set_name) && $group->check_member($member)) {
+	return $group->delete_member($member);
+    }
+    return;
 }
 
 sub ipset_check_set_type {
